feat(AENS): Fix BC for all of AENS commands. Add additional option `extendPointers` for `update` command which allow to extend pointer instead of overwriting them Fix test's
