+++ bin/commands/aens.js
@@ -31,7 +31,7 @@ async function preClaim (walletPath, domain, options) {
 
   try {
     // Validate `name`(check if `name` end on `.test`)
-    validateName(domain)
+    // validateName(domain)
 
     // Get `keyPair` by `walletPath`, decrypt using password and initialize `Ae` client with this `keyPair`
     const client = await initClientByWalletFile(walletPath, options)
@@ -66,7 +66,7 @@ async function claim (walletPath, domain, salt, options) {
   const { ttl, fee, nonce, waitMined, json, nameFee } = options
   try {
     // Validate `name`
-    validateName(domain)
+    // validateName(domain)
 
     // Get `keyPair` by `walletPath`, decrypt using password and initialize `Ae` client with this `keyPair`
     const client = await initClientByWalletFile(walletPath, options)
@@ -98,12 +98,13 @@ async function claim (walletPath, domain, salt, options) {
 }
 
 // ##Update `name` function
-async function updateName (walletPath, domain, address, options) {
-  const { ttl, fee, nonce, waitMined, json, nameTtl, clientTtl } = options
+async function updateName (walletPath, domain, addresses, options) {
+  const { ttl, fee, nonce, waitMined, json, nameTtl, clientTtl, extendPointers = false } = options
 
   try {
     // Validate `address`
-    if (!isAddressValid(address)) throw new Error(`Address "${address}" is not valid`)
+    const invalidAddresses = addresses.filter(address => !isAddressValid(address))
+    if (invalidAddresses.length) throw new Error(`Addresses "[${invalidAddresses}]" is not valid`)
     // Validate `name`
     validateName(domain)
     // Get `keyPair` by `walletPath`, decrypt using password and initialize `Ae` client with this `keyPair`
@@ -118,7 +119,7 @@ async function updateName (walletPath, domain, address, options) {
       }
 
       // Create `updateName` transaction
-      const updateTx = await client.aensUpdate(name.id, address, { ttl, fee, nonce, waitMined, nameTtl, clientTtl })
+      const updateTx = await client.aensUpdate(domain, addresses, { ttl, fee, nonce, waitMined, nameTtl, clientTtl, extendPointers })
       if (waitMined) {
         printTransaction(
           updateTx,
@@ -192,7 +193,7 @@ async function revokeName (walletPath, domain, options) {
       }
 
       // Create `revokeName` transaction
-      const revokeTx = await client.aensRevoke(name.id, { ttl, fee, nonce, waitMined })
+      const revokeTx = await client.aensRevoke(domain, { ttl, fee, nonce, waitMined })
       if (waitMined) {
         printTransaction(
           revokeTx,
@@ -271,7 +272,7 @@ async function fullClaim (walletPath, domain, options) {
       nonce += 1
       const claim = await preclaim.claim({ nonce, ttl, fee, nameFee })
       nonce += 1
-      const updateTx = await claim.update(await client.address(), { nonce, ttl, fee, nameTtl, clientTtl })
+      const updateTx = await claim.update([await client.address()], { nonce, ttl, fee, nameTtl, clientTtl })
       nonce += 1
 
       printTransaction(
