+++ test/cli/tx.js
@@ -17,7 +17,7 @@
 
 import { describe, it } from 'mocha'
 
-import { configure, BaseAe, execute, parseBlock, ready } from './index'
+import { configure, BaseAe, execute, parseBlock, ready, randomString } from './index'
 import { decodeBase64Check, generateKeyPair } from '@aeternity/aepp-sdk/es/utils/crypto'
 import MemoryAccount from '@aeternity/aepp-sdk/es/account/memory'
 import fs from 'fs'
@@ -30,16 +30,6 @@ const testContract = `contract Identity =
 
 function randomName (length = 18, namespace = '.chain') {
   return randomString(length).toLowerCase() + namespace
-}
-
-function randomString (len, charSet) {
-  charSet = charSet || 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789'
-  let randomString = ''
-  for (let i = 0; i < len; i++) {
-    const randomPoz = Math.floor(Math.random() * charSet.length)
-    randomString += charSet.substring(randomPoz, randomPoz + 1)
-  }
-  return randomString
 }
 
 async function signAndPost (tx, assert) {
